+++ spec/controllers/concerns/authentication_spec.rb
@@ -84,11 +84,6 @@ RSpec.describe Veri::Authentication do
     it "logs in the user" do
       expect { subject }.to change(controller, :current_user).from(nil).to(user)
     end
-
-    it "runs after_login callback" do
-      allow(controller).to receive(:after_login).and_return("foo")
-      expect(subject).to eq("foo")
-    end
   end
 
   describe "#log_out" do
@@ -104,16 +99,6 @@ RSpec.describe Veri::Authentication do
 
     it "logs out the user" do
       expect { subject }.to change(user.veri_sessions, :count).from(1).to(0)
-    end
-
-    it "runs after_logout callback" do
-      allow(controller).to receive(:after_logout).and_return("foo")
-      expect(subject).to eq("foo")
-    end
-
-    it "does nothing if no user is logged in" do
-      controller.log_out
-      expect { controller.log_out }.not_to raise_error
     end
   end
 
@@ -159,19 +144,71 @@ RSpec.describe Veri::Authentication do
     end
   end
 
+  describe "#shapeshifter?" do
+    subject { controller.shapeshifter? }
+
+    let(:controller) { DummyController.new }
+
+    before { controller.request = ActionDispatch::TestRequest.create }
+
+    context "when there is no current session" do
+      it { is_expected.to be false }
+    end
+
+    context "when user has shapeshifted" do
+      let(:user) { User.create! }
+
+      before do
+        controller.log_in(user)
+        controller.current_session.shapeshift(user)
+      end
+
+      it { is_expected.to be true }
+    end
+
+    context "when user has not shapeshifted" do
+      let(:user) { User.create! }
+
+      before { controller.log_in(user) }
+
+      it { is_expected.to be false }
+    end
+
+    context "when user has reverted to true identity" do
+      let(:user) { User.create! }
+
+      before do
+        controller.log_in(user)
+        controller.current_session.shapeshift(user)
+        controller.current_session.revert_to_true_identity
+      end
+
+      it { is_expected.to be false }
+    end
+  end
+
   describe "helper methods" do
     let(:controller) { DummyController.new }
     let(:view) { controller.view_context }
     let(:user) { User.create! }
+    let(:session) { Veri::Session.new }
 
-    before { allow(controller).to receive_messages(current_user: user, logged_in?: true) }
+    before { allow(controller).to receive_messages(current_user: user, logged_in?: true, shapeshifter?: true, current_session: session) }
 
     it "provides current_user helper method" do
-      expect(view.current_user).to eq(user)
+      expect(view.current_user).to be user
     end
 
     it "provides logged_in? helper method" do
       expect(view.logged_in?).to be true
+    end
+
+    it "provides shapeshifter? helper method" do
+      expect(view.shapeshifter?).to be true
+    end
+
+    it "provides current_session helper method" do
+      expect(view.current_session).to be session
     end
   end
 end